requirements.txt
# ── Web Framework ──────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ── Database & ORM ─────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0

# ── Validation & Serialization ─────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# ── Authentication & Security ──────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.17

# ── HTTP & Utilities ───────────────────────────────────────────
httpx==0.28.0
python-dotenv==1.0.1

# ── Rate Limiting ──────────────────────────────────────────────
slowapi==0.1.9

# ── Dev / Tooling ──────────────────────────────────────────────
rich==13.9.4
